refactor: split common file into several files
