+++ jba/src/visualization/common/utils.py
@@ -4,6 +4,8 @@ import pandas as pd
 
 from jba.src.models.edu_columns import EduColumnName
 
+ALL_CHOICE_OPTIONS = 'All'
+
 
 def get_edu_name_columns(df: pd.DataFrame) -> List[str]:
     df_columns = df.columns.tolist()
