+++ jba/src/visualization/pages/2_Group_tests_timeline.py
@@ -6,7 +6,7 @@ import streamlit as st
 from core.src.utils.df_utils import read_df
 from jba.src.models.edu_columns import EduColumnName, EduTaskType
 from jba.src.models.edu_logs import TestDataField, TestResult
-from jba.src.visualization.common import (
+from jba.src.visualization.common.filters import (
     filter_post_correct_submissions,
     filter_by_task,
     filter_by_group,
